+++ freopen.c
@@ -31,6 +31,7 @@ hidden FILE *next_freopen(const char *path, const char *mode, FILE *stream)
 
 FILE *freopen(const char *path, const char *mode, FILE *stream)
 {
+	const int errno_save = errno;
 	char buf[PATH_MAX];
 	FILE *ret = NULL;
 	ssize_t siz;
@@ -40,8 +41,8 @@ FILE *freopen(const char *path, const char *mode, FILE *stream)
 		goto exit;
 
 	ret = next_freopen(buf, mode, stream);
-	if (ret != NULL)
-		__setfd(fileno(ret), buf);
+	if (ret != NULL && __setfd(fileno(ret), buf))
+		errno = errno_save;
 
 exit:
 	__debug("%s(path: '%s' -> '%s', mode: '%s', ...) -> %p\n", __func__,