+++ open.c
@@ -37,6 +37,7 @@ hidden int next_open(const char *path, int oflags, mode_t mode)
 
 int open(const char *path, int oflags, ...)
 {
+	const int errno_save = errno;
 	mode_t oldmode = 0, mode = 0;
 	int atflags = 0, ret = -1;
 	char buf[PATH_MAX];
@@ -65,8 +66,8 @@ int open(const char *path, int oflags, ...)
 
 	ret = next_open(buf, oflags, mode);
 	__set_mode(buf, oldmode, mode);
-	if (ret >= 0)
-		__setfd(ret, buf);
+	if (ret >= 0 && __setfd(ret, buf))
+		errno = errno_save;
 
 exit:
 	__debug("%s(path: '%s' -> '%s', oflags: 0%o, mode: 0%03o -> 0%03o) -> %i\n",