+++ CHANGELOG.md
@@ -13,6 +13,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fix the `ld-iamroot.so(8)` option `--cwd`
+- Fix the alteration of the `errno` error in the functions `__open_2()`,
+  `__openat_2()`, `close()`, `dup()`, `dup2()`, `dup3()`, `fopen()`,
+  `freopen()`, `open()` and `openat()`,
 
 ## [v22] - 2024-06-04